requirements.txt
python-dotenv

lumibot
timedelta
alpaca-trade-api

langchain
langchain_community
gpt4all